+++ src/audio_output_helper.cpp
@@ -3,7 +3,7 @@
 // WebRTC
 #include <rtc_base/logging.h>
 
-#if defined(SORA_CPP_SDK_IOS)
+#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_VISIONOS)
 #include "sora/mac/mac_audio_output_helper.h"
 #endif
 
@@ -23,7 +23,7 @@ class DummyAudioOutputHelper : public AudioOutputHelperInterface {
 
 std::unique_ptr<AudioOutputHelperInterface> CreateAudioOutputHelper(
     AudioChangeRouteObserver* observer) {
-#if defined(SORA_CPP_SDK_IOS)
+#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_VISIONOS)
   return std::make_unique<MacAudioOutputHelper>(observer);
 #else
   return std::make_unique<DummyAudioOutputHelper>();