+++ src/sora_video_codec.cpp
@@ -8,7 +8,8 @@
 #include "sora/java_context.h"
 #include "sora/open_h264_video_codec.h"
 
-#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS)
+#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS) || \
+    defined(SORA_CPP_SDK_VISIONOS)
 #include "sora/mac/mac_video_factory.h"
 #elif defined(SORA_CPP_SDK_ANDROID)
 #include "sora/android/android_video_factory.h"
@@ -282,7 +283,8 @@ VideoCodecCapability GetVideoCodecCapability(
   {
     auto& engine =
         cap.engines.emplace_back(VideoCodecImplementation::kInternal);
-#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS)
+#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS) || \
+    defined(SORA_CPP_SDK_VISIONOS)
     auto internal_encoder_factory = CreateMacVideoEncoderFactory();
     auto internal_decoder_factory = CreateMacVideoDecoderFactory();
 #elif defined(SORA_CPP_SDK_ANDROID)