+++ src/sora_video_codec_factory.cpp
@@ -84,7 +84,8 @@ std::optional<SoraVideoCodecFactory> CreateVideoCodecFactory(
                                 VideoCodecImplementation::kInternal;
                    }) != preference.codecs.end()) {
     builtin_encoder_factory = webrtc::CreateBuiltinVideoEncoderFactory();
-#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS)
+#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS) || \
+    defined(SORA_CPP_SDK_VISIONOS)
     builtin_encoder_factory = CreateMacVideoEncoderFactory();
 #elif defined(SORA_CPP_SDK_ANDROID)
     builtin_encoder_factory =
@@ -102,7 +103,8 @@ std::optional<SoraVideoCodecFactory> CreateVideoCodecFactory(
                             *codec.decoder ==
                                 VideoCodecImplementation::kInternal;
                    }) != preference.codecs.end()) {
-#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS)
+#if defined(SORA_CPP_SDK_IOS) || defined(SORA_CPP_SDK_MACOS) || \
+    defined(SORA_CPP_SDK_VISIONOS)
     builtin_decoder_factory = CreateMacVideoDecoderFactory();
 #elif defined(SORA_CPP_SDK_ANDROID)
     builtin_decoder_factory =